Debug and fixed regex failure in test_votd.
The reason for failure was because of the regex pattern used. The older
pattern could not find books that were prepended with a number, such as
"1Th".

Also, enable verbose mode in make command, i.e. `make verbose`.

Old: ref.ly/[A-Z,a-z]+[1-9]+.[1-9,-]+;

New: ref.ly/[A-Z,a-z,1-9]+[1-9]+.[1-9,-]+;

For example, scraping the following raw html failed with the old pattern:

<a href='https://twitter.com/share?via=Logos&amp;url=https://ref.ly/1Th5.6;asv%201901&amp;text="so%20then%20let%20us%20not%20sleep,%20as%20do%20the%20rest,%20but%20let%20us%20watch%20and%20be%20sober."' id="twitter-link" rel="noopener noreferrer" target="_blank"></a>

Updates:
- Makefile
- src/pybiblia/pybiblia.py

Excludes:
- setup distribution